ps_sni: cast send() length to int for Windows -Wconversion fix
On Windows (MinGW), send() takes an int length; len - sent yields
size_t, triggering -Werror=conversion. Cast to int is safe since the
loop guard ensures the remaining length fits.

+++ src/openvpn/ps_sni.c
@@ -728,7 +728,7 @@ sni_passthrough_send_client_hello(socket_descriptor_t sd, const char *sni,
     sent = 0;
     while (sent < (ssize_t)len)
     {
-        ssize_t n = send(sd, (const char *)(buf + sent), len - sent, MSG_NOSIGNAL);
+        ssize_t n = send(sd, (const char *)(buf + sent), (int)(len - sent), MSG_NOSIGNAL);
         if (n <= 0)
         {
             msg(D_LINK_ERRORS | M_ERRNO, "--sni-passthrough-hostname: send() failed");
